sdk-requirements.txt
loguru~=0.6.0
PyYAML~=6.0
pandas~=1.5.3
python-dotenv~=0.21.1
omegaconf~=2.3.0
tiktoken~=0.2.0
openai~=0.26.5